requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
schedule==1.1.0
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
